First Review for backup module
- Now use a template file
- tool.php is now more or less apeform like.
- Manual include of libs and files mostly removed
- No more manual access check needed
- FTAN implemented
- Cleaned up folder Structure
- Uses "no_page" now.
- Added return to admin tools button.
- No more risk that an attacker attacks the backup-sql.php

Using an include isn't the perfect solution(can't be reused),
but it certainly will do for now.
